==> R/01_GSE_load.qmd <==
---
title: "GSE_load"
format: html
editor: visual
---

```{r setup}
own_path <- "/Users/mathildebechgaard/"
knitr::opts_knit$set(root.dir = paste0(own_path, "group_04_project/data/"))
```

## Getting libraries

```{r, message = FALSE, warning = FALSE}
library(GEOquery)
library(limma)
library(tidyverse)
library(edgeR)
```

## Loading data

```{r}
gse_data <- getGEO(GEO = "GSE39833")[[1]]

exp_data <- exprs(gse_data) #extracting expression values (Limma)

gpl <- Table(getGEO("GPL14767"))
```

## Tidying data

```{r}
gse_data_frame <- as.data.frame(exp_data) |> 
  rownames_to_column(var = "ID")

gpl <- gpl |> 
  mutate(ID = as.character(ID))

mirna_exp <- gse_data_frame |> 
  left_join(gpl, by = "ID") |> 
  select( ,-1) |> 
  relocate(where(is.numeric), .after = where(is.character))

metadata <- gse_data@phenoData@data
```

## Saving datasets

```{r}
saveRDS(gse_data_frame, file = paste0(own_path, "group_04_project/data/gse_data_load.RDs"))

exp_data <- as.data.frame(exp_data)
saveRDS(exp_data, file = paste0(own_path, "group_04_project/data/gse_expressed_data.RDs"))

saveRDS(gpl, file = paste0(own_path, "group_04_project/data/gse_chosen_meta_data.RDs"))

saveRDS(mirna_exp, file = paste0(own_path, "group_04_project/data/gse_all_data.RDs"))

saveRDS(metadata, file = paste0(own_path, "group_04_project/data/gse_all_meta_data.RDs"))
```
