==> R/02_GSE_clean.qmd <==
---
title: "GSE_new_df"
format: html
editor: visual
---

```{r setup}
own_path <- "/Users/mathildebechgaard/"
knitr::opts_knit$set(root.dir = paste0(own_path, "group_04_project/data/"))
```

## Loading library

```{r, message = FALSE, warning = FALSE}
library(tidyverse)
```

## Creating new data frame

```{r}
patient_stage <- readRDS(paste0(own_path, ("group_04_project/data/gse_all_meta_data.RDs"))) |> 
  select(source_name_ch1, characteristics_ch1.2, characteristics_ch1.3, characteristics_ch1.4) |> 
  mutate(group = if_else(characteristics_ch1.4 == "tnm stages: normal", "control", "cancer")) |> 
  mutate(characteristics_ch1.3 = as.numeric(str_replace(characteristics_ch1.3, "age: ", ""))) |> 
  mutate(characteristics_ch1.2 = str_replace(characteristics_ch1.2, "gender: ", "")) |> 
  mutate(characteristics_ch1.4 = str_replace(characteristics_ch1.4, "tnm stages: ", "")) |> 
  dplyr::rename("gender" = characteristics_ch1.2,
         "age" = characteristics_ch1.3,
         "cancer_stage" = characteristics_ch1.4)
```

## Saving file

```{r}
patient_stage <- as.data.frame(patient_stage)
saveRDS(patient_stage, file = paste0(own_path, "group_04_project/data/gse_clean.RDs"))
```
