==> R/03_TCGA_augment.qmd <==
---
title: "TCGA_augment"
format: html
editor: visual
---

## Set Working directory

```{r setup, message = FALSE, warning = FALSE}
own_path <- "/Users/mathildebechgaard"
knitr::opts_knit$set(root.dir = paste0(own_path, "/group_04_project/data/"))
source(paste0(own_path, "/group_04_project/project_functions.R"))
```

## Load Libraries

```{r, message = FALSE, warning = FALSE}
library(edgeR)
library(tibble)
library(statmod)
library(tidyverse)
library(stringr)
```

## Fetch datasets

```{r}
#miRNA
miRNA_data <- readRDS(paste0(own_path, "/group_04_project/data/miRNA_data_clean.RDS"))
miRNA_metadata <- readRDS(paste0(own_path, "/group_04_project/data/miRNA_metadata_clean.RDS"))

#mRNA
mRNA_data <- readRDS(paste0(own_path, "/group_04_project/data/mRNA_data_clean.RDS"))
mRNA_metadata <- readRDS(paste0(own_path, "/group_04_project/data/mRNA_metadata_clean.RDS"))
```

## miRNA data pre-processing

Creating a meta dataset that will indicate the tissue type for each of the TCGA IDs:

1\) Extracting column names from the created `miRNA_data`.\
2) Identifying the tissue type by IDs numbering (e.g., 01A = cancer, 11A = normal tissue) - and defining the result as a metadata for the patients IDs.

```{r}
column_names <- colnames(miRNA_data)[-1] 
IDs <- tibble(TCGA_ID = column_names) # e.g., TCGA-F4-6854-01A
```

```{r}
metadata_id_patients_miRNA <- IDs |>
  set_tissue_status() |>
  arrange(Status)
```

Log2 transformation, and handling NAs. If all values are missing the function replaces the entire column with **`NA`**. If there is at least one non-missing value in a column, it removes the missing values from that column using **`na.omit().`**

```{r}
miRNA_log <- miRNA_data |>
  mutate(across(where(is.numeric), ~log2(. + 1)))

miRNA_log <- as.data.frame(miRNA_log) |>
  group_by(miRNA_ID) |>
  summarise(across(everything(), ~if(all(is.na(.))) NA else na.omit(.)))
```

## mRNA data pre-processing

Setting "Status" from the existing "definition" column in `mRNA_metadata`.

```{r}
mRNA_metadata <- mRNA_metadata |>
  mutate(Status = str_extract(definition, "\\w+$")) |>
  arrange(Status) |>
  select(-definition)
```

Creating metadata for the patients IDs.

```{r}
metadata_id_patients_mRNA <- mRNA_metadata |>
  select(sample, Status)
```

Log2 transformation and making mRNA dataset similar to the miRNA dataset to apply the further functions with edgeR processing.

```{r}
mRNA_log <- mRNA_data |>
  mutate(across(where(is.numeric), ~log2(. + 1))) |>
  mutate(genes = row.names(mRNA_data)) |>
  as_tibble() |>
  select(genes, everything())

mRNA_log <- as.data.frame(mRNA_log) |>
  group_by(genes) |>
  summarise(across(everything(), ~if(all(is.na(.))) NA else na.omit(.)))
```

## edgeR

#### Creating normalized dataset

```{r}
mRNA_y <- process_data(mRNA_log)
miRNA_y <- process_data(miRNA_log)
```

#### Function to run edgeR differential expression analysis

In the next code chunk, a function for an edgeR processing is used. The function could be found in the "project_functions.R" filr.

Firstly, the design matrix is shaped, that will compare groups "Cancer" and "Normal" that are indicated in a column "Status" in metadata dataframes.

Secondly, the following actions are performed:

1\) **Estimation of a common dispersion (across all genes) and tagwise (gene-specific) dispersions**. It is recommended to do both, for a proper capturing of the variability in gene expression data. It leads to more reliable and biologically meaningful results when performing the differential gene expression analysis.

2\) **Fitting of a quasi-likelihood (QL) negative binomial model** to the data. It is designed for count data, which is encountered in RNA-seq experiments.

3\) Performance of a **quasi-likelihood F-test** on the fitted QL model. It allows to assess the significance of differences in gene expression between two conditions (cancer and normal tissues). F-test generates adjusted p-values to assess the significance of the changes in genes expression.

4\) **`topTags`** is used to **extract the top genes**, based on statistical significance from the results of a differential expression analysis. It takes the output of `glmQLFTest` and returns a list of top-ranked genes, based on the calculated values that indicate the significance.

### Saving statistics datasets

Saving the statistics information about the differential expression into `top_genes_stat` and gene names into `top_genes` to filter them in the datasets for the further analysis.

```{r}
# mRNA data
mRNA_edgeR <- run_edgeR_analysis(mRNA_y, "~ Status", metadata_id_patients_mRNA, mRNA_data)
mRNA_top_genes_stat <- mRNA_edgeR$table 
mRNA_top_genes <- mRNA_top_genes_stat$genes

# miRNA data
# setting the factors to correctly compare the values
metadata_id_patients_miRNA$Status <- factor(metadata_id_patients_miRNA$Status, 
                                           levels = c("normal", "cancer"))

miRNA_edgeR <- run_edgeR_analysis(miRNA_y, "~ Status", metadata_id_patients_miRNA, miRNA_data)
miRNA_top_genes_stat <- miRNA_edgeR$table 
miRNA_top_genes <- miRNA_top_genes_stat$miRNA_ID
```

#### Pivoting and filtering the datasets

From the existing log datasets, only top deferentially expressed genes, identified by edgeR, are chosen.

```{r}
mRNA_log_long <- pivot_and_filter(mRNA_log, genes, mRNA_top_genes)
miRNA_log_long <- pivot_and_filter(miRNA_log, miRNA_ID, miRNA_top_genes)
```

-   mRNA

```{r}
mRNA_metadata <- mRNA_metadata |>
  dplyr::rename("TCGA_ID" = barcode)
```

Selecting columns of interest for the analysis.

```{r}
mRNA_metadata <- mRNA_metadata |>
  select(patient, Status, ajcc_pathologic_stage, alcohol_history, gender, vital_status, age_at_index, days_to_death)
```

-   miRNA

Selecting columns of interest for the analysis.

```{r}
miRNA_metadata <- miRNA_metadata |>
  dplyr::rename("TCGA_ID" = case_submitter_id)
```

```{r}
miRNA_metadata <- miRNA_metadata |>
  select(TCGA_ID, ajcc_pathologic_stage,
         gender, age_at_index, days_to_death) |>
  set_tissue_status()
```

## Gene names as row names for expression datasets

```{r}
row_names <- miRNA_log$miRNA_ID
miRNA_log <- miRNA_log[,-1]
rownames(miRNA_log) <- row_names

row_names <- mRNA_log$genes
mRNA_log <- mRNA_log[,-1]
rownames(mRNA_log) <- row_names
```

## Saving datasets for analysing

```{r}
saveRDS(miRNA_log, file = paste0(own_path, "/group_04_project/data/miRNA_expression.RDs"))
saveRDS(miRNA_log_long, file = paste0(own_path, "/group_04_project/data/miRNA_aug.RDs"))
saveRDS(miRNA_top_genes_stat, file = paste0(own_path, "/group_04_project/data/miRNA_top_genes_stat.RDs"))
saveRDS(miRNA_metadata, file = paste0(own_path, "/group_04_project/data/miRNA_metadata_aug.RDs"))

saveRDS(mRNA_log, file = paste0(own_path, "/group_04_project/data/mRNA_expression.RDS"))
saveRDS(mRNA_log_long, file = paste0(own_path, "/group_04_project/data/mRNA_aug.RDS"))
saveRDS(mRNA_top_genes_stat, file = paste0(own_path, "/group_04_project/data/mRNA_top_genes_stat.RDs"))
saveRDS(mRNA_metadata, file = paste0(own_path, "/group_04_project/data/mRNA_metadata_aug.RDs"))
```

## Clean environment

```{r}
rm(list = ls())
```
