==> R/04_GSE_describe.qmd <==
---
title: "Describe"
format: html
editor: visual
---

```{r setup}
own_path <- "/Users/mathildebechgaard/"
knitr::opts_knit$set(root.dir = paste0(own_path, "group_04_project/data/"))
```

## Loading libraries

```{r, message = FALSE, warning = FALSE}
library(tidyverse)
library(ggplot2)
```

## Load Data

```{r}
patient_stage <- readRDS(paste0(own_path, ("group_04_project/data/gse_clean.RDs")))
```

## Histogram

```{r}
gse_plot1 <- patient_stage |>  
  ggplot(mapping = aes(x = cancer_stage,
         fill = gender)) +
  geom_bar(position = "dodge",
           colour = "black",
           alpha = 0.5) +
  labs(x = "Cancer stage", y = "Occurences") +
  theme(panel.grid.major.x = element_blank()) 
gse_plot1
```

```{r}
gse_plot2 <- ggplot(patient_stage, aes(x = cancer_stage, fill = cancer_stage)) +
  geom_bar() +
  labs(x = "The stages of cancer ", y = "Frequency", title = "Frequency of the different states of cancer") +
  theme(axis.text.x = element_blank())
gse_plot2
```

## Boxplot

```{r, warning = FALSE}
gse_plot3 <- ggplot(patient_stage, aes(x = gender, y = age, fill = gender)) + geom_violin() +
  labs(x = "", y = "Age", title = "miRNA from the GEO data") +
  guides(fill = FALSE) +
  facet_wrap(~ "gender", ncol = 2, scales = "free_x") +
  geom_boxplot(width=0.1)
gse_plot3
```

## Saving the plots

```{r}
ggsave(file.path("group_04_project/results/images/", "occurences_gender.png"), gse_plot1, width = 10, height = 6, units = "in")
ggsave(file.path("group_04_project/results/images/", "cancer_frequency.png"), gse_plot2, width = 10, height = 6, units = "in")
ggsave(file.path("group_04_project/results/images/", "gse_gender_distribution.png"), gse_plot3, width = 10, height = 6, units = "in")
```
