==> R/04_TCGA_describe.qmd <==
---
title: "TCGA_describe"
format: html
editor: visual
---

## Set Working directory

```{r setup}
own_path <- "/Users/mathildebechgaard"
knitr::opts_knit$set(root.dir = paste0(own_path, "/group_04_project/data/"))
```

## Load Libraries

```{r, message = FALSE, warning = FALSE}
library(tidyverse)
library(patchwork)
```

## Fetch datasets

```{r}
miRNA_metadata <- readRDS(paste0(own_path, "/group_04_project/data/miRNA_metadata_aug.RDS"))
mRNA_metadata <- readRDS(paste0(own_path, "/group_04_project/data/mRNA_metadata_aug.RDS"))
```

## Experiment description

#### Number og Cancer vs. Normal Samples

```{r, warning = FALSE}
my_colors <- c("Normal" = "cyan3", "Tumor" = "cyan4", "normal" = "cyan3", "cancer" = "cyan4")

plot1 <- ggplot(miRNA_metadata, aes(x = Status, fill = Status)) +
  geom_bar() +
  labs(x = "", y = "Frequency", title = "miRNA") +
  theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust = 1)) +
  guides(fill = FALSE) +
  scale_fill_manual(values = my_colors) +
  coord_cartesian(ylim = c(0, 75)) +
  facet_wrap(~ "Diagnosis", ncol = 2, scales = "free_x")

plot2 <- ggplot(mRNA_metadata, aes(x = Status, fill = Status)) +
  geom_bar() +
  labs(x = "", y = "", title = "mRNA") +
  theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust = 1)) +
  scale_fill_manual(values = my_colors) +
  coord_cartesian(ylim = c(0, 75)) +
  facet_wrap(~ "Diagnosis", ncol = 2, scales = "free_x")

cancer_vs_normal <- plot1 + plot2 + plot_layout(ncol = 2) +
  plot_annotation(title = "Frequency of Cases by Category", theme = theme(plot.title = element_text(hjust = 0.5)))

cancer_vs_normal

```

#### Gender distribution

```{r}
# Plot for miRNA_metadata with adjusted fill colors and y-axis limit
plot1 <- ggplot(miRNA_metadata, aes(x = gender, y = age_at_index, fill = gender)) + geom_violin() +
  labs(x = "", y = "Age", title = "miRNA") +
  guides(fill = FALSE) +
  facet_wrap(~ "Gender", ncol = 2, scales = "free_x") +
  geom_boxplot(width=0.1)

# Plot for mRNA_metadata with adjusted fill colors and y-axis limit
plot2 <- ggplot(mRNA_metadata, aes(x = gender, y = age_at_index, fill = gender)) + geom_violin() +
  labs(x = "", y = "", title = "mRNA") +
  facet_wrap(~ "Gender", ncol = 2, scales = "free_x") +
  geom_boxplot(width=0.1)

# Combine plots using patchwork and add a main title
gender_distribution <- plot1 + plot2 + plot_layout(ncol = 2) +
  plot_annotation(title = "Distribution of Age by Gender", theme = theme(plot.title = element_text(hjust = 0.5)))

gender_distribution
```

#### Cancer stages

```{r}
cancer_stage <- vctrs::vec_c(miRNA_metadata$ajcc_pathologic_stage, mRNA_metadata$ajcc_pathologic_stage)
cancer_stage_df <- data.frame(stage = unlist(cancer_stage))

cancer_stages <- ggplot(cancer_stage_df, aes(x = stage, fill = stage)) +
  geom_bar() +
  labs(x = "AJCC Pathologic Stage", y = "Frequency", title = "Frequency of AJCC Pathologic Stages by Status") +
  theme(axis.text.x = element_blank())

cancer_stages
```

## Save plots

```{r}
ggsave(file.path("../results/images/", "cancer_vs_normal.png"), cancer_vs_normal, width = 10, height = 6, units = "in")
ggsave(file.path("../results/images/", "gender_distribution.png"), gender_distribution, width = 10, height = 6, units = "in")
ggsave(file.path("../results/images/", "cancer_stages.png"), cancer_stages, width = 10, height = 6, units = "in")
```

## Clean environment

```{r}
rm(list = ls())
```
