==> R/05_TCGA_miRNA_analysis.qmd <==
---
title: "miRNA_analysis"
format: html
editor: visual
---

## Set Working directory

```{r setup, message = FALSE, warning = FALSE}
own_path <- "/Users/mathildebechgaard"
knitr::opts_knit$set(root.dir = paste0(own_path, "/group_04_project/data/"))
source(paste0(own_path, "/group_04_project/project_functions.R"))
```

## Load Libraries

```{r, message = FALSE, warning = FALSE}
library(tidyverse)
library(ggrepel)
```

## Fetch dataset

```{r}
miRNA_log_long <- readRDS(paste0(own_path, "/group_04_project/data/miRNA_aug.RDS"))
top_genes_stat <- readRDS(paste0(own_path, "/group_04_project/data/miRNA_top_genes_stat.RDS"))
miRNA_meta <- readRDS(paste0(own_path, "/group_04_project/data/mRNA_metadata_clean.RDS"))
```

Subsetting top 15 "log Fold Changed" genes for a better visualisation:

```{r}
top_genes_stat_subset <- top_genes_stat |>
  arrange(desc(abs(logFC))) |>
  head(n = 15)

top_genes <- top_genes_stat_subset$miRNA_ID
```

Uniting Stages in one group to facet the plots based on them later on.

```{r}
miRNA_meta <- miRNA_meta |>
 mutate(Stage = case_when(
    ajcc_pathologic_stage == "Stage I" | ajcc_pathologic_stage == "Stage IA" ~ "Stage I",
    ajcc_pathologic_stage == "Stage II" | ajcc_pathologic_stage == "Stage IIA" |
    ajcc_pathologic_stage == "Stage IIB" ~ "Stage II",
    ajcc_pathologic_stage == "Stage III" | ajcc_pathologic_stage == "Stage IIIA" |
    ajcc_pathologic_stage == "Stage IIIB" | ajcc_pathologic_stage == "Stage IIIC" ~ "Stage III",
    ajcc_pathologic_stage == "Stage IV" | ajcc_pathologic_stage == "Stage IVB" ~ "Stage IV"
  )) |>
  select(-ajcc_pathologic_stage) |> 
  tibble() |>
  select(-sample_submitter_id) |>
  dplyr::rename(TCGA_ID = "sample")
```

## Comparison of the expression between the stages

```{r}
miRNA_box <- miRNA_meta |> 
  select(TCGA_ID, Stage) |> 
  right_join(miRNA_log_long, by = 'TCGA_ID') |>
  set_tissue_status() |>
  ggplot(aes(x = Stage,
         y = log_reads,
         col = Status)) +
  geom_boxplot() +
  scale_color_manual(values = c("cancer" = "#f47c6c", 
                                "normal" = "#13b3ba")) + 
  labs(title = "Boxplot comparison of total miRNA expression between four cancer stages", color = "Tissue Type",
       x = "Cancer stage",
       y = "Log2 expression") +
  theme(plot.title = element_text(size=20),
        axis.title = element_text(size=15), 
        legend.text = element_text(size=15),
        legend.title = element_text(size=20))
```

## General comparison of the expression between cancer and normal tissues

-   Ideally, the samples should be taken from cancer and normal tissues of the same patients to get a proper overview of the expression difference. In our case, not all the samples had this comparison. Therefore, to make the conclusions based on the plots more precise, one should properly collect the expression data.

-   From the graph below, it can be seen that, in general, miRNAs expression is higher in cancer tissues.

```{r}
miRNA_expression <- miRNA_log_long |>
  filter(miRNA_ID %in% top_genes) |>
  set_tissue_status() |>
  ggplot(aes(x = miRNA_ID,
             y = log_reads,
             col = Status)) +
  geom_point() +
  labs(title = "Comparison of expression of 15 top differentially expressed miRNAs \nbetween cancer and normal TCGA samples", color = "Tissue Type",
       x = "miRNA IDs",
       y = "Log2 expression") +
  scale_color_manual(values = c("cancer" = "#f47c6c", 
                                "normal" = "#13b3ba")) +
  theme(axis.text.x = element_text(
    angle = 45,
    hjust = 1),
    plot.title = element_text(
      size = 12, 
      angle = 0)) +
  theme(plot.title = element_text(size=20),
        axis.title = element_text(size=15), 
        legend.text = element_text(size=15),
        legend.title = element_text(size=20))
```

## Volcano plot

-   In the following volcano plot, statistically significant values are highlighted in blue (the p-value is flipped, so it becomes more significant with the higher values). Positive LFC indicates higher expression in the cancer group compared to the normal tissues, that is in accordance with the plots above.

```{r}
# stating the significance based on the adjusted pvalues (FDR corrects for multiple comparisons)
top_genes_stat <- top_genes_stat |>
    mutate(
      significance = case_when(
        FDR > 0.05 ~ "no",
        FDR <= 0.05 ~ "yes"
      )
    )
```

```{r}
top_genes_stat <- top_genes_stat |>
  mutate(
    regulation = case_when(
      logFC > 1 ~ "Upregulated",
      logFC < -1 ~ "Downregulated",
      TRUE ~ "Non-significant"
    )
  )
```

```{r}
mirna_vol <- top_genes_stat |> 
  ggplot(aes(x = logFC,
             y = -log10(FDR),
             col = regulation)) +
  geom_point(alpha = 0.5) +
  scale_color_manual(values = c("Upregulated" = "red", 
                                "Downregulated" = "blue", 
                                "Non-significant" = "gray"),
                     name = "Type of genes regulation") +
  geom_hline(yintercept = -log10(0.05), linetype = "dashed", color = "black") +
  geom_vline(xintercept = c(-1, 1), linetype = "dashed", color = "black") +
  labs(x = "Log Fold Change",
       y = "-log10(FDR)",
       title = "Top Differentially Expressed miRNAs in TCGA-COAD Samples") +
  theme(plot.title = element_text(hjust = 0.5)) +
  geom_text_repel(data = top_genes_stat_subset,
                  aes(label = miRNA_ID),
                      color = "black") +
  theme(plot.title = element_text(size=20),
        axis.title = element_text(size=15), 
        legend.text = element_text(size=15),
        legend.title = element_text(size=20))
```

## Save plots

```{r}
ggsave(file.path(file = paste0(own_path, "/group_04_project/results/images/mirna_boxplot.png")), plot = miRNA_box, width = 14, height = 8, unit = "in")

ggsave(file.path(file = paste0(own_path, "/group_04_project/results/images/mirna_expression.png")), plot = miRNA_expression, width = 14, height = 8, unit = "in")

ggsave(file.path(file = paste0(own_path, "/group_04_project/results/images/mirna_vol.png")), plot = mirna_vol, width = 14, height = 8, unit = "in")
```

## Clean environment

```{r}
rm(list = ls())
```
