==> R/07_Combined_heatmap.qmd <==
---
title: "Combined_Heatmap"
format: html
editor: visual
---

## Set Working directory

```{r setup}
own_path <- "/Users/mathildebechgaard"
knitr::opts_knit$set(root.dir = paste0(own_path, "/group_04_project/data/"))
```

## Load Libraries

```{r, message = FALSE, warning = FALSE}
library(tidyverse)
library(ggplot2)
library(reshape2)
```

## Fetch dataset

```{r}
gse_miRNA <- readRDS(paste0(own_path, "/group_04_project/data/gse_top_20.rds"))
TCGA_miRNA <- readRDS(paste0(own_path, "/group_04_project/data/miRNA_top_expression.rds"))
TCGA_mRNA <- readRDS(paste0(own_path, "/group_04_project/data/mRNA_top_expression.rds"))
```

## Transform data for visualization

```{r, message = FALSE, warning = FALSE}
GSE_miRNA <- gse_miRNA |>
  mutate(across(where(is.numeric), ~log2(. + 1)))

TCGA_miRNA <- TCGA_miRNA |>
  mutate(across(where(is.numeric), ~log2(. + 1)))
rownames(TCGA_miRNA) <- TCGA_miRNA[, 1]
TCGA_miRNA <- TCGA_miRNA[, -1]
TCGA_miRNA <- TCGA_miRNA[, colSums(is.na(TCGA_miRNA)) != nrow(TCGA_miRNA), drop = FALSE]

TCGA_mRNA <- as.data.frame(TCGA_mRNA) |>
  mutate(across(where(is.numeric), ~log2(. + 1)))
rownames(TCGA_mRNA) <- TCGA_mRNA[, 1]
TCGA_mRNA <- TCGA_mRNA[, -1]
```

## Transpose dataframes to fit geom_tile

```{r}
GSE_miRNA <- t(GSE_miRNA)
TCGA_miRNA <- t(TCGA_miRNA)
TCGA_mRNA <- t(TCGA_mRNA)
```

## miRNA Heatmap

```{r}
GSE_miRNA_heatmap <- ggplot(data = melt(GSE_miRNA), aes(x = Var2, y = Var1, fill = value)) +
  geom_tile() +
  labs(x = "Genes", y = "Patients", title = "Heatmap of Expression Values") +
  theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1),
        axis.text.y = element_blank())
GSE_miRNA_heatmap + scale_fill_gradient(limits = c(-10, 15))
```

## miRNA Heatmap

```{r}
TCGA_miRNA_heatmap <- ggplot(data = melt(TCGA_miRNA), aes(x = Var2, y = Var1, fill = value)) +
  geom_tile() +
  labs(x = "Genes", y = "Patients", title = "Heatmap of Expression Values") +
  theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1),
        axis.text.y = element_blank())
TCGA_miRNA_heatmap + scale_fill_gradient(limits = c(-10, 15))
```

## mRNA Heatmap

```{r}
TCGA_mRNA_heatmap <- ggplot(data = melt(TCGA_mRNA), aes(x = Var2, y = Var1, fill = value)) +
  geom_tile() +
  labs(x = "Genes", y = "Patients", title = "Heatmap of Expression Values") +
  theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1),
        axis.text.y = element_blank())
TCGA_mRNA_heatmap + scale_fill_gradient(limits = c(-10, 15))
```

## Save plots

```{r}
ggsave(file.path("../results/images/", "GSE_miRNA_heatmap.png"), GSE_miRNA_heatmap, width = 10, height = 6, units = "in")
ggsave(file.path("../results/images/", "TCGA_miRNA_heatmap.png"), TCGA_miRNA_heatmap, width = 10, height = 6, units = "in")
ggsave(file.path("../results/images/", "TCGA_mRNA_heatmap.png"), TCGA_mRNA_heatmap, width = 10, height = 6, units = "in")
```
